fix: dark mode and readme css
