+++ packages/ui-components/package.json
@@ -33,7 +33,6 @@
     "country-flag-icons": "1.5.5",
     "dayjs": "1.11.7",
     "dompurify": "2.4.5",
-    "github-markdown-css": "5.2.0",
     "highlight.js": "11.7.0",
     "history": "4.10.1",
     "i18next": "20.6.1",
